+++ fvr_engine-client/src/sparse_quad_grid.rs
@@ -27,7 +27,7 @@ where
 {
     pub fn new(width: u32, height: u32) -> Result<Self> {
         let vertices = GridMap::new(width, height);
-        let indices = generate_indices(width * height);
+        let indices = generate_indices((width * height) as usize);
         let vertex_data = Vec::new();
 
         let mut vbo = 0;
